+++ hpOneView/resources/networking/fcoe_networks.py
@@ -44,7 +44,8 @@ class FcoeNetworks(object):
     DEFAULT_VALUES = {
         '200': {"type": "fcoe-network"},
         '300': {"type": "fcoe-networkV300"},
-        '500': {"type": "fcoe-networkV300"}
+        '500': {"type": "fcoe-networkV300"},
+        '600': {"type": "fcoe-networkV4"}
     }
 
     def __init__(self, con):