+++ tests/unit/resources/servers/test_enclosure_groups.py
@@ -72,15 +72,15 @@ class EnclosureGroupsTest(unittest.TestCase):
     def test_get_all_called_once(self, mock_get_all):
         filter = 'name=TestName'
         sort = 'name:ascending'
+        scope_uris = 'rest/scopes/cd237b60-09e2-45c4-829e-082e318a6d2a'
+        self.client.get_all(2, 500, filter, sort, scope_uris)
 
-        self.client.get_all(2, 500, filter, sort)
-
-        mock_get_all.assert_called_once_with(2, 500, filter=filter, sort=sort)
+        mock_get_all.assert_called_once_with(2, 500, filter=filter, sort=sort, scope_uris=scope_uris)
 
     @mock.patch.object(ResourceClient, 'get_all')
     def test_get_all_called_once_with_default(self, mock_get_all):
         self.client.get_all()
-        mock_get_all.assert_called_once_with(0, -1, filter='', sort='')
+        mock_get_all.assert_called_once_with(0, -1, filter='', sort='', scope_uris='')
 
     @mock.patch.object(ResourceClient, 'get')
     def test_get_by_id_called_once(self, mock_get):