+++ hpOneView/resources/networking/logical_switch_groups.py
@@ -47,7 +47,8 @@ class LogicalSwitchGroups(object):
     DEFAULT_VALUES = {
         '200': {"type": "logical-switch-group"},
         '300': {"type": "logical-switch-groupV300"},
-        '500': {"type": "logical-switch-groupV300"}
+        '500': {"type": "logical-switch-groupV300"},
+        '600': {"type": "logical-switch-groupV4"}
     }
 
     def __init__(self, con):