+++ tests/unit/resources/servers/test_enclosures.py
@@ -40,16 +40,17 @@ class EnclosuresTest(TestCase):
     def test_get_all_called_once(self, mock_get_all):
         filter = 'name=TestName'
         sort = 'name:ascending'
+        scope_uris = 'rest/scopes/cd237b60-09e2-45c4-829e-082e318a6d2a'
 
-        self._enclosures.get_all(2, 500, filter, sort)
+        self._enclosures.get_all(2, 500, filter, sort, scope_uris)
 
-        mock_get_all.assert_called_once_with(2, 500, filter=filter, sort=sort)
+        mock_get_all.assert_called_once_with(2, 500, filter=filter, sort=sort, scope_uris=scope_uris)
 
     @mock.patch.object(ResourceClient, 'get_all')
     def test_get_all_called_once_with_default_values(self, mock_get_all):
         self._enclosures.get_all()
 
-        mock_get_all.assert_called_once_with(0, -1, filter='', sort='')
+        mock_get_all.assert_called_once_with(0, -1, filter='', sort='', scope_uris='')
 
     @mock.patch.object(ResourceClient, 'get_by')
     def test_get_by_called_once(self, mock_get_by):
@@ -240,3 +241,48 @@ class EnclosuresTest(TestCase):
 
         mock_get.assert_called_once_with('/rest/enclosures/09USE7335NW3',
                                          fields=None, filter=None, refresh=False, view=None)
+
+    @mock.patch.object(ResourceClient, 'create')
+    def test_generate_csr(self, mock_create):
+        bay_number = 1
+        id_enclosure = 'ad28cf21-8b15-4f92-bdcf-51cb2042db32'
+        uri_rest_call = '/rest/enclosures/ad28cf21-8b15-4f92-bdcf-51cb2042db32/https/certificaterequest?bayNumber=%d' % (bay_number)
+        csr_data = {
+            'type': 'CertificateDtoV2',
+            'organization': 'Acme Corp.',
+            'organizationalUnit': 'IT',
+            'locality': 'Townburgh',
+            'state': 'Mississippi',
+            'country': 'US',
+            'email': 'admin@example.com'
+        }
+        headers = {'Content-Type': 'application/json'}
+
+        self._enclosures.generate_csr(csr_data, id_enclosure, bay_number=bay_number)
+
+        mock_create.assert_called_once_with(csr_data, uri=uri_rest_call, custom_headers=headers)
+
+    @mock.patch.object(ResourceClient, 'get')
+    def test_get_csr(self, mock_get):
+        bay_number = 1
+        id_enclosure = 'ad28cf21-8b15-4f92-bdcf-51cb2042db32'
+        uri_rest_call = '/rest/enclosures/ad28cf21-8b15-4f92-bdcf-51cb2042db32/https/certificaterequest?bayNumber=%d' % (bay_number)
+
+        self._enclosures.get_csr(id_enclosure, bay_number=bay_number)
+
+        mock_get.assert_called_once_with(uri_rest_call)
+
+    @mock.patch.object(ResourceClient, 'update')
+    def test_import_certificate(self, mock_update):
+        bay_number = 1
+        id_enclosure = 'ad28cf21-8b15-4f92-bdcf-51cb2042db32'
+        uri_rest_call = '/rest/enclosures/ad28cf21-8b15-4f92-bdcf-51cb2042db32/https/certificaterequest?bayNumber=%d' % (bay_number)
+        certificate_data = {
+            'type': 'CertificateDataV2',
+            'base64Data': '-----BEGIN CERTIFICATE----- encoded data here -----END CERTIFICATE-----'
+        }
+        headers = {'Content-Type': 'application/json'}
+
+        self._enclosures.import_certificate(certificate_data, id_enclosure, bay_number=bay_number)
+
+        mock_update.assert_called_once_with(certificate_data, uri=uri_rest_call, custom_headers=headers)