+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 #### Features supported with current release:
 - Connection template
 - FC network
+- FCoE Networks
 - Interconnect type
 - Logical interconnect group
 - SAS logical interconnect group