+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 #### Features supported with current release:
 - Connection template
 - FC network
-- FCoE Networks
+- FCoE network
 - Interconnect type
 - Logical interconnect group
 - SAS logical interconnect group