+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Interconnect type
 - Logical interconnect group
 - SAS logical interconnect group
+- SAS logical interconnect
 
 # 4.4.0
 #### Notes