+++ CHANGELOG.md
@@ -13,13 +13,13 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Interconnect type
 - Internal link set
 - Logical interconnect group
+- Logical Switch Groups
 - Logical switch
 - SAS interconnect type
 - SAS logical interconnect group
 - SAS logical interconnect
 - Switch type
 - Uplink set
-
 
 # 4.4.0
 #### Notes