+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 #### Features supported with current release:
 - Connection template
 - Enclosure group
+- Enclosure
 - Ethernet network
 - FC network
 - FCoE network