+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - SAS logical interconnect group
 - SAS logical interconnect
 - Switch type
+- Tasks
 - Uplink set
 
 # 4.4.0