+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - SAS logical interconnect group
 - Server hardware
 - Switch type
+- Tasks
 - Uplink set
 
 # 4.4.0