+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - SAS logical interconnect
 - SAS logical interconnect group
 - Server hardware
+- Switch
 - Switch type
 - Tasks
 - Uplink set