+++ CHANGELOG.md
@@ -22,9 +22,9 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - SAS logical interconnect
 - SAS logical interconnect group
 - Server hardware
+- Server profile
 - Server profile template
 - Switch
-- Server profile
 - Switch type
 - Tasks
 - Uplink set