+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 4.5.0 (Unreleased)
+# 4.5.0 
 #### Notes
 Added the capability to set a connection timeout when connecting to the HPE OneView Appliance
 
@@ -27,7 +27,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Server profile template
 - Switch
 - Switch type
-- Tasks
+- Task
 - Uplink set
 
 # 4.4.0